+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # DeepNest
 
+## 0.1.1
+
+### Include DeepNest in native Ruby classes
+
+  - Add helper namespace to use self in recursive methods (i.e. can use `structure.deep_dup` instead of `DeepNest.deep_dup(structure)`).
+  - Add patch file to allow users to monkey patch recursive methods in native Hash and Array Ruby classes.
+  - Add patch spec.
+
 ## 0.1.0
 
 ### Initial Release